+++ src/main/java/edu/ksu/canvas/impl/AssignmentOverrideImpl.java
@@ -29,14 +29,14 @@ public class AssignmentOverrideImpl extends BaseImpl<AssignmentOverride, Assignm
     }
     
     @Override
-    public List<AssignmentOverride> listAssignmentOverrides(String courseId, Integer assignmentId) throws IOException {
+    public List<AssignmentOverride> listAssignmentOverrides(String courseId, Long assignmentId) throws IOException {
         LOG.debug("Retrieving a list of assignment overrides in course " + courseId + " for assignment " + assignmentId);
         String url = buildCanvasUrl("courses/" + courseId + "/assignments/" + assignmentId + "/overrides", Collections.emptyMap());
         return getListFromCanvas(url);
     }
 
     @Override
-    public Optional<AssignmentOverride> getAssignmentOverride(String courseId, Integer assignmentId, Integer overrideId) throws IOException {
+    public Optional<AssignmentOverride> getAssignmentOverride(String courseId, Long assignmentId, Long overrideId) throws IOException {
         LOG.debug("Retrieving an assignment override in course " + courseId + " for assignment " + assignmentId);
         String url = buildCanvasUrl("courses/" + courseId + "/assignments/" + assignmentId + "/overrides/" + overrideId, Collections.emptyMap());
         Response response = canvasMessenger.getSingleResponseFromCanvas(oauthToken, url);
