+++ src/main/java/edu/ksu/canvas/model/Conversation.java
@@ -19,13 +19,13 @@ public class Conversation extends BaseCanvasModel implements Serializable {
     private Date lastMessageAt;
     private String lastAuthoredMessage;
     private Date lastAuthoredMessageAt;
-    private Integer messageCount;
+    private Long messageCount;
     private Boolean subscribed;
     @SerializedName("private") //can't name a variable "private" in java
     private Boolean isPrivate;
     private Boolean starred;
     private List<ConversationFlags> properties;
-    private List<Integer> audience;
+    private List<Long> audience;
     private Object audienceContexts; //TODO: Refine this object type
     private String avatarUrl;
     private List<MessageParticipant> participants;
@@ -46,7 +46,7 @@ public class Conversation extends BaseCanvasModel implements Serializable {
         private List<Object> forwardedMessages;
         private List<String> attachments;
         private Object mediaComment;
-        private List<Integer> participatingUserIds;
+        private List<Long> participatingUserIds;
 
         public Long getId() {
             return id;
@@ -112,11 +112,11 @@ public class Conversation extends BaseCanvasModel implements Serializable {
             this.mediaComment = mediaComment;
         }
 
-        public List<Integer> getParticipatingUserIds() {
+        public List<Long> getParticipatingUserIds() {
             return participatingUserIds;
         }
 
-        public void setParticipatingUserIds(List<Integer> participatingUserIds) {
+        public void setParticipatingUserIds(List<Long> participatingUserIds) {
             this.participatingUserIds = participatingUserIds;
         }
     }
@@ -227,11 +227,11 @@ public class Conversation extends BaseCanvasModel implements Serializable {
         this.lastAuthoredMessageAt = lastAuthoredMessageAt;
     }
 
-    public Integer getMessageCount() {
+    public Long getMessageCount() {
         return messageCount;
     }
 
-    public void setMessageCount(Integer messageCount) {
+    public void setMessageCount(Long messageCount) {
         this.messageCount = messageCount;
     }
 
@@ -267,11 +267,11 @@ public class Conversation extends BaseCanvasModel implements Serializable {
         this.properties = properties;
     }
 
-    public List<Integer> getAudience() {
+    public List<Long> getAudience() {
         return audience;
     }
 
-    public void setAudience(List<Integer> audience) {
+    public void setAudience(List<Long> audience) {
         this.audience = audience;
     }
